pyproject.toml
[project]
name = "aliasai-validator"
version = "0.1.0"
description = "ERC-8004 Staking Validation System Complete Test Toolkit"
authors = [
    { name = "AliasAI", email = "dev@aliasai.com" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["erc8004", "ethereum", "validation", "staking", "blockchain", "web3"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "web3>=6.15.0",
    "eth-account>=0.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/AliasAI/aliasai-validator"
Documentation = "https://github.com/AliasAI/aliasai-validator#readme"
Repository = "https://github.com/AliasAI/aliasai-validator"
Issues = "https://github.com/AliasAI/aliasai-validator/issues"

[project.scripts]
erc8004-test = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
